- Объединил состояние игры в один объект - Внес в класс состояния все функции, манипулирующие только состоянием игры - Вынес объявление методов класса из конструктора
