+++ js/utils/routeToNextPage.js
@@ -3,7 +3,6 @@ import {QuestionType} from '../consts';
 import renderPage from './renderPage';
 import mainLevelArtistPage from '../components/mainLevelArtist/mainLevelArtistPage';
 import mainLevelGenrePage from '../components/mainLevelGenre/mainLevelGenrePage';
-import getUserScore from './getUserScore';
 import getGameResultMessage from './getGameResultMessage';
 import resultWinPage from '../components/resultWin/resultWinPage';
 import resultLoosePage from '../components/resultLoose/resultLoosePage';
@@ -15,27 +14,30 @@ import resultTimeoutPage from '../components/resultTimeout/resultTimeoutPage';
  * @param {Object} gameState Состояние игры
  * @param {number} gameState.lives Количество жизней игрока
  * @param {number} gameState.time Оставшееся время игры
- * @param {Object} gameData Данные игры
- * @param {boolean} gameData.hasNextQuestion Признак того, что есть следующий вопрос
+ * @param {boolean} gameState.hasNextQuestion Признак того, что есть следующий вопрос
+ * @param {Object} gameState.currentQuestion Текущий вопрос игры
+ * @param {string} gameState.currentQuestion.type Тип текущего вопроса игры
+ * @param {Object} gameState.userScore Объект счета игрока
+ * @param {number} gameState.userScore.totalScore Общий счет игрока
  */
-export default function routeToNextPage(gameState, gameData) {
+export default function routeToNextPage(gameState) {
   if (gameState.time <= 0) {
     renderPage(resultTimeoutPage(getGameResultMessage({time: gameState.time})));
   } else if (gameState.lives <= 0) {
     renderPage(resultLoosePage(getGameResultMessage({lives: gameState.lives})));
-  } else if (gameData.hasNextQuestion) {
-    gameData = gameData.iterateQuestion();
+  } else if (gameState.hasNextQuestion) {
+    gameState = gameState.iterateQuestion();
 
-    if (gameData.currentQuestion.type === QuestionType.ARTIST) {
-      renderPage(mainLevelArtistPage(gameState, gameData));
-    } else if (gameData.currentQuestion.type === QuestionType.GENRE) {
-      renderPage(mainLevelGenrePage(gameState, gameData));
+    if (gameState.currentQuestion.type === QuestionType.ARTIST) {
+      renderPage(mainLevelArtistPage(gameState));
+    } else if (gameState.currentQuestion.type === QuestionType.GENRE) {
+      renderPage(mainLevelGenrePage(gameState));
     }
   } else {
-    const userScore = getUserScore(gameState.lives, gameData);
+    const userScore = gameState.userScore;
     // TODO: вторым параметром передавать результаты других игроков
     const resultMessage = getGameResultMessage({
-      score: userScore.score,
+      score: userScore.totalScore,
       lives: gameState.lives,
       time: gameState.time
     }, []);
